% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_divisions.R
\name{add_divisions}
\alias{add_divisions}
\title{Add Conference and Division Column}
\usage{
add_divisions(df)
}
\arguments{
\item{df}{pbp data}
}
\value{
df with added columns
}
\description{
Add Conference and Division Column
}
