% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_epa.R
\name{adjust_epa}
\alias{adjust_epa}
\title{Title}
\usage{
adjust_epa(pbp)
}
\arguments{
\item{pbp}{play by play data from nflfastR}
}
\value{
df object
}
\description{
based on https://dtai.cs.kuleuven.be/events/MLSA18/papers/pelechrinis_mlsa18.pdf
}
