% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_games.R
\name{double_games}
\alias{double_games}
\title{Double Games}
\usage{
double_games(df, load_data = FALSE)
}
\arguments{
\item{df}{of you want to load the data this can be ignored}

\item{load_data}{if you want to load the data from github. The default is FALSE}
}
\value{
df
}
\description{
Helper function that takes input where each game has one row with teams as \code{away_team} and \code{home_team} and
returns with each game having two rows with teams listed as \code{team} and \code{opp}
}
