% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pbp_data.R
\name{get_pbp_data}
\alias{get_pbp_data}
\title{Load PBP Data}
\usage{
get_pbp_data(seasons = c(2021))
}
\arguments{
\item{seasons}{what years do you want to load}
}
\value{

}
\description{
Load PBP Data
}
\examples{
get_pbp_data(seasons = 2021)
}
