% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonic_mean.R
\name{harmonic_mean}
\alias{harmonic_mean}
\title{Calculate Harmonic Mean}
\usage{
harmonic_mean(x, na.rm = TRUE)
}
\arguments{
\item{x}{numerical values}

\item{na.rm}{remove NA values}
}
\value{
harmonic mean value
}
\description{
Calculate Harmonic Mean
}
\examples{
faithful[['waiting']] \%>\% harmonic_mean()
}
