% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/height_to_cm.R
\name{height_to_cm}
\alias{height_to_cm}
\title{Height to cm}
\usage{
height_to_cm(x)
}
\arguments{
\item{x}{height character value}
}
\value{
numerical value in cm
}
\description{
convert character string of height e.g. "6-4" to a numerical
value in cm.   Note: it will break if the height is above 10feet.
}
\examples{
height_to_cm("6-4")
}
