% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ngs.R
\name{load_ngs}
\alias{load_ngs}
\title{Load NGS}
\usage{
load_ngs(
  stat_type = c("passing", "receiving", "rushing"),
  year = c(2021),
  weekly = TRUE,
  regular_season = TRUE,
  join_roster = FALSE,
  pbp_data = pbp_data
)
}
\arguments{
\item{stat_type}{either passing, receiving or rushing}

\item{year}{the year you want the stats from. Starts from 2016. Default is current year.}

\item{weekly}{if you want the weekly or season stats. Default is yes.}

\item{regular_season}{do you want to exclude postseason? Default is yes.}

\item{join_roster}{do you want to join to pbp data? Default is no.}

\item{pbp_data}{if you want to join with pbp data, define what your pbp is called.}
}
\value{

}
\description{
Load NGS
}
\examples{
load_ngs(stat_type = "rushing")
}
