% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_see_difference.R
\name{rolling_sse_difference}
\alias{rolling_sse_difference}
\title{Rolling SSE Difference}
\usage{
rolling_sse_difference(x, n = 2)
}
\arguments{
\item{x}{tot.withinss column}

\item{n}{amount to lag}
}
\value{
numerical values
}
\description{
Rolling SSE Difference
}
