% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_epa.R
\name{weekly_epa}
\alias{weekly_epa}
\title{Calculate Weekly EPA}
\usage{
weekly_epa(pbp)
}
\arguments{
\item{pbp}{play by play data from nflfastR}
}
\value{
Returns weekly epa/play for each team.
}
\description{
Calculate Weekly EPA
}
